requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.5
distro==1.9.0
elastic-transport==8.13.0
elasticsearch==8.13.0
fastapi==0.110.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.49
langchain-elasticsearch==0.1.3
langchain-openai==0.1.5
langchain-text-splitters==0.0.1
langsmith==0.1.53
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.25.1
orjson==3.10.2
outcome==1.3.0.post0
packaging==23.2
pdf2image==1.17.0
pillow==10.3.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.3
pypdf==4.2.0
PySocks==1.7.1
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
result==0.16.1
selenium==4.20.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
webdriver-manager==4.0.1
wsproto==1.2.0
yarl==1.9.4
